Run TestPionE2ESimpleED25519 only on Go 1.13
Conditionally build TestPionE2ESimpleED25519 in e2e_v113_test.go.
Once Go 1.12 is deprecated, move this test to e2e_test.go.
